+++ src/data/emacs-mode/agda2-mode.el
@@ -521,11 +521,11 @@ The error message directs the user to the *agda2* buffer."
        (eq (agda2-process-status) 'run)))
 
 (defun agda2-send-command (restart &rest args)
-  "Send a command to the Agda2 asyncronous process.  Saves the buffer
-and sends the list of strings ARGS to Agda2.  If RESTART is non-nil
-and the Agda2 process is not running, or the *agda2* buffer does not
-exist, then an attempt is made to restart the process."
-  (save-buffer)
+  "Send a command to the Agda process.
+Sends the list of strings ARGS to the process. If RESTART is
+non-nil and the process is not running, or the *agda2*
+buffer does not exist, then an attempt is made to restart the
+process."
   (when (and restart (not (agda2-running-p)))
     ;; Try restarting automatically, but only once, in case there is
     ;; some major problem.
@@ -545,10 +545,11 @@ exist, then an attempt is made to restart the process."
     (error "Another command is currently in progress
 \(if a command has been aborted you may want to restart Agda)")))
 
-(defun agda2-go (responses-expected highlight do-abort &rest args)
+(defun agda2-go (save responses-expected highlight do-abort &rest args)
   "Executes commands in the Agda2 interpreter.
-Sends the list of strings ARGS to the Agda2 interpreter, waits for
-output and executes the responses, if any. If no responses are
+Sends the list of strings ARGS to the Agda2 interpreter, waits
+for output and executes the responses, if any. If SAVE is
+'save, then the buffer is saved first. If no responses are
 received, and RESPONSES-EXPECTED is non-nil, then an error is
 raised. If HIGHLIGHT is non-nil, then the buffer's syntax
 highlighting may be updated."
@@ -561,6 +562,8 @@ highlighting may be updated."
         agda2-responses               0
         agda2-output-chunk-incomplete (agda2-queue-empty)
         agda2-file-buffer             (current-buffer))
+
+  (when (equal save 'save) (save-buffer))
 
   (apply 'agda2-send-command
          'restart
@@ -724,7 +727,7 @@ This function should be used in `first-change-hook'."
     (message "\"%s\" has been modified. Interrupting highlighting."
              (buffer-name (current-buffer)))))
 
-(defun agda2-goal-cmd (cmd &optional want ask &rest args)
+(defun agda2-goal-cmd (cmd save &optional want ask &rest args)
   "Reads input from goal or minibuffer and sends command to Agda.
 
 An error is raised if point is not in a goal.
@@ -747,6 +750,9 @@ The user input is computed as follows:
 If the user input is not taken from the goal, then an empty goal
 range is given.
 
+If SAVE is 'save, then the buffer is saved just before the
+command is sent to Agda (if it is sent).
+
 An error is raised if no responses are received."
   (multiple-value-bind (o g) (agda2-goal-at (point))
     (unless g (error "For this command, please place the cursor in a goal"))
@@ -758,7 +764,7 @@ An error is raised if no responses are received."
                   (or ask (string-match "\\`\\s *\\'" txt)))
              (setq txt (read-string (concat want ": ") nil nil txt t)))
             (t (setq input-from-goal t)))
-      (apply 'agda2-go t input-from-goal t cmd
+      (apply 'agda2-go save t input-from-goal t cmd
              (format "%d" g)
              (if input-from-goal (agda2-goal-Range o) (agda2-mkRange nil))
              (agda2-string-quote txt) args))))
@@ -780,7 +786,7 @@ An error is raised if no responses are received."
 (defun agda2-load ()
   "Load current buffer."
   (interactive)
-  (agda2-go t t t "Cmd_load"
+  (agda2-go 'save t t t "Cmd_load"
             (agda2-string-quote (buffer-file-name))
             (agda2-list-quote agda2-program-args)
             ))
@@ -814,7 +820,7 @@ The variable `agda2-backend' determines which backend is used."
                                          'inherit-input-method))
                        (t agda2-backend))))
     (when (equal backend "") (error "No backend chosen"))
-    (agda2-go t t t "Cmd_compile"
+    (agda2-go 'save t t t "Cmd_compile"
               backend
               (agda2-string-quote (buffer-file-name))
               (agda2-list-quote agda2-program-args)
@@ -822,7 +828,7 @@ The variable `agda2-backend' determines which backend is used."
 
 (defun agda2-give()
   "Give to the goal at point the expression in it" (interactive)
-  (agda2-goal-cmd "Cmd_give" "expression to give"))
+  (agda2-goal-cmd "Cmd_give" 'save "expression to give"))
 
 (defun agda2-give-action (old-g paren)
   "Update the goal OLD-G with the expression in it."
@@ -848,18 +854,18 @@ goal is replaced by the constructor applied to a suitable number
 of new goals."
   (interactive "P")
   (if pmlambda
-      (agda2-goal-cmd "Cmd_refine_or_intro True" 'goal)
-    (agda2-goal-cmd "Cmd_refine_or_intro False" 'goal)))
+      (agda2-goal-cmd "Cmd_refine_or_intro True" 'save 'goal)
+    (agda2-goal-cmd "Cmd_refine_or_intro False" 'save 'goal)))
 
 (defun agda2-auto ()
  "Simple proof search" (interactive)
- (agda2-goal-cmd "Cmd_auto" 'goal))
+ (agda2-goal-cmd "Cmd_auto" 'save 'goal))
 
 (defun agda2-make-case ()
   "Refine the pattern variables given in the goal.
 Assumes that <clause> = {!<variables>!} is on one line."
   (interactive)
-  (agda2-goal-cmd "Cmd_make_case" "pattern variables to case (empty for split on result)"))
+  (agda2-goal-cmd "Cmd_make_case" 'save "pattern variables to case (empty for split on result)"))
 
 (defun agda2-make-case-action (newcls)
   "Replace the line at point with new clauses NEWCLS and reload."
@@ -1126,11 +1132,11 @@ is inserted, and point is placed before this text."
 
 (defun agda2-show-goals()
   "Show all goals." (interactive)
-  (agda2-go t t t "Cmd_metas"))
+  (agda2-go nil t t t "Cmd_metas"))
 
 (defun agda2-show-constraints()
   "Show constraints." (interactive)
-  (agda2-go t t t "Cmd_constraints"))
+  (agda2-go nil t t t "Cmd_constraints"))
 
 (defun agda2-remove-annotations ()
   "Removes buffer annotations (overlays and text properties)."
@@ -1176,7 +1182,7 @@ COMMENT is used to build the function's comment. The function
 NAME takes a prefix argument which tells whether it should
 normalise types or not when running CMD (through
 `agda2-goal-cmd'; WANT is used as `agda2-goal-cmd's WANT
-argument)."
+argument, and nil as its SAVE argument)."
   (let ((eval (make-symbol "eval")))
   `(defun ,name (&optional prefix)
      ,(concat comment ".
@@ -1198,15 +1204,15 @@ The form of the result depends on the prefix argument:
                         ((equal prefix '(4)) "Instantiated")
                         ("Normalised"))))
        (agda2-goal-cmd (concat ,cmd " " ,eval)
-                       ,want)))))
+                       nil ,want)))))
 
 (defmacro agda2-maybe-normalised-toplevel (name comment cmd prompt)
   "This macro constructs a function NAME which runs CMD.
 COMMENT is used to build the function's comments. The function
 NAME takes a prefix argument which tells whether it should
 normalise types or not when running CMD (through
-`agda2-go' t nil t; the string PROMPT is used as the goal command
-prompt)."
+`agda2-go' nil t nil t; the string PROMPT is used as the goal
+command prompt)."
   (let ((eval (make-symbol "eval")))
     `(defun ,name (prefix expr)
        ,(concat comment ".
@@ -1227,7 +1233,7 @@ The form of the result depends on the prefix argument:
        (let ((,eval (cond ((equal prefix nil) "Simplified")
                           ((equal prefix '(4)) "Instantiated")
                           ("Normalised"))))
-         (agda2-go t nil t
+         (agda2-go nil t nil t
                    (concat ,cmd " " ,eval " "
                            (agda2-string-quote expr)))))))
 
@@ -1236,7 +1242,7 @@ The form of the result depends on the prefix argument:
 COMMENT is used to build the function's comments. The function
 NAME takes a prefix argument which tells whether it should
 normalise types or not when running CMD (through
-`agda2-go' t nil t;)."
+`agda2-go' nil t nil t;)."
   (let ((eval (make-symbol "eval")))
     `(defun ,name (prefix)
        ,(concat comment ".
@@ -1257,7 +1263,7 @@ The form of the result depends on the prefix argument:
        (let ((,eval (cond ((equal prefix nil) "AsIs")
                           ((equal prefix '(4)) "Simplified")
                           ("Normalised"))))
-         (agda2-go t nil t
+         (agda2-go nil t nil t
                    (concat ,cmd " " ,eval " "
                            ))))))
 
@@ -1293,12 +1299,12 @@ top-level scope."
 (defun agda2-why-in-scope ()
   "Explain why something is in scope in a goal."
   (interactive)
-  (agda2-goal-cmd "Cmd_why_in_scope" "Name"))
+  (agda2-goal-cmd "Cmd_why_in_scope" nil "Name"))
 
 (defun agda2-why-in-scope-toplevel (name)
   "Explain why something is in scope at the top level."
   (interactive "MName: ")
-  (agda2-go t nil t
+  (agda2-go nil t nil t
             "Cmd_why_in_scope_toplevel"
             (agda2-string-quote name)))
 
@@ -1416,7 +1422,7 @@ computation."
                       (cond ((equal arg nil) " DefaultCompute")
                             ((equal arg '(4)) " IgnoreAbstract")
                             (" UseShowInstance")))))
-    (agda2-goal-cmd cmd "expression to normalise")))
+    (agda2-goal-cmd cmd nil "expression to normalise")))
 
 (defun agda2-compute-normalised-toplevel (expr &optional arg)
   "Compute the normal form of the given expression.
@@ -1434,7 +1440,7 @@ computation."
                      (cond ((equal arg nil) " DefaultCompute")
                             ((equal arg '(4)) " IgnoreAbstract")
                             (" UseShowInstance")) " ")))
-    (agda2-go t nil t (concat cmd (agda2-string-quote expr)))))
+    (agda2-go nil t nil t (concat cmd (agda2-string-quote expr)))))
 
 (defun agda2-compute-normalised-maybe-toplevel ()
   "Compute the normal form of the given expression.
@@ -1455,7 +1461,7 @@ computation."
 (defun agda2-display-program-version ()
   "Display version of Agda"
   (interactive)
-  (agda2-go nil nil t "Cmd_show_version"))
+  (agda2-go nil nil nil t "Cmd_show_version"))
 
 ;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
 ;;;;
@@ -1464,7 +1470,7 @@ computation."
   "Loads precomputed syntax highlighting info for the current buffer.
 Only if the buffer is unmodified, and only if there is anything to load."
  (unless (buffer-modified-p)
-   (agda2-go nil t t
+   (agda2-go nil nil t t
              "Cmd_load_highlighting_info"
              (agda2-string-quote (buffer-file-name)))))
 
@@ -1604,7 +1610,7 @@ text properties."
         (delete-region p (+ p 2)))
         ;; Update highlighting
         (if (and (not (equal new-txt 'paren)) (not (equal new-txt 'no-paren)))
-            (apply 'agda2-go t t nil "Cmd_highlight"
+            (apply 'agda2-go 'save t t nil "Cmd_highlight"
               (format "%d" old-g)
               (agda2-mkRange `(,p ,(- q 2)))
               (agda2-string-quote new-txt) nil))
@@ -1871,10 +1877,11 @@ configuration and the selected window are not changed."
 With prefix argument, turn on display of implicit arguments if
 the argument is a positive number, otherwise turn it off."
   (interactive "P")
-  (cond ((eq arg nil)       (agda2-go t t t "ToggleImplicitArgs"))
-        ((and (numberp arg)
-              (> arg 0))    (agda2-go t t t "ShowImplicitArgs" "True"))
-        (t                  (agda2-go t t t "ShowImplicitArgs" "False"))))
+  (cond
+   ((eq arg nil)       (agda2-go nil t t t "ToggleImplicitArgs"))
+   ((and (numberp arg)
+         (> arg 0))    (agda2-go nil t t t "ShowImplicitArgs" "True"))
+   (t                  (agda2-go nil t t t "ShowImplicitArgs" "False"))))
 
 ;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
 ;;;;